requirements.txt
cffi==1.16.0
cryptography==42.0.8
oracledb==2.2.1
pycparser==2.22
python-dotenv==1.0.1
